pyproject.toml
[tool.poetry]
name = "online-courses"
version = "0.1.0"
description = ""
authors = ["Roman.Iastremskyi <iastremskyi010397@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.6"
djangorestframework = "^3.15.2"
psycopg2-binary = "^2.9.10"
django-filter = "^25.1"
djangorestframework-simplejwt = "^5.4.0"


[tool.poetry.group.dev.dependencies]
pillow = "^11.1.0"
python-dotenv = "^1.0.1"
coverage = "^7.6.12"


[tool.poetry.group.lint.dependencies]
isort = "^6.0.0"
black = "^25.1.0"
flake8 = "^7.1.2"
mypy = "^1.15.0"

[tool.black]
line-length = 119
exclude = '''
(
  /(
    | \.git
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | dist
  )/
)
'''

[tool.mypy]
disallow_untyped_defs = true
no_implicit_optional = true
warn_return_any = true
exclude = '.venv'

[tool.isort]
line_length = 119

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
